Don't panic on errors
I ran into a hard to debug error where Terraform was erroring with:

```
rpc error: code = Unavailable desc = transport is closing
```

This turned out to be because my BK token didn't have the righ
permissions so the SDK was returning an error and the provider was
panicing because it was logged with `log.Fatalf`. This logs using
`log.Printf` and adds diagnostics where some where missing.