Add a diagnostic if the manifest can't be found
Previously if you typo'd the manifest name Terraform would blow up
because we hadn't appended anything to `diag` so it was `nil` and we
were returning `nil` for the the artifact. This ended up in the
provider plugin segfaulting.

Now we get a diagnostic error message:

```
Error: Could not find manifest

Could not find manifest does_not_exist.json for build number 470011
```