Scroll through paginated buildkite artifacts to find manifest
The buildkite API paginates the list artifacts for job endpoint[1] so
when there are jobs that we output a lot of artifacts for, the manifest
we're searching for may not be included in the first page. This adds
scrolling through the pages until we either find the manifest or run out
of pages.

Originally I had wrapped up getting _all_ the artifacts then just
iterating through those. It's a little less complicated but didn't make
full use of pagination. Now we only scroll until the manifest is found
and return early.