Allow specifying a fallback manifest
We want to be able to deploy something when the manifest hasn't been
built yet (or won't be built). For example, because the build failed
or we're deploying before it has completed to get out a time sensitive
change - these are all break glass measures but we need to be able to
support them. This adds a `fallback_manifest` attribute to the
provider which can be optionally provided. If provided and the
expected manifest does not exist, then the fallback manifest will be
used. This fallback can be disabled altogether by setting the
`STILE_MANIFEST_NO_FALLBACK` for normal operation where we do want to
fail to catch issues with these manifests early on.

Note: The manifest not existing is handled separately to other errors
trying to get the manifest (e.g. authentication error or the manifest
being invalid JSON), these will still be errors.